python/requirements.txt
flask==3.0.0
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
